ARMv7: GICv2 driver can manage GICv1 with security extension
Some SoCs integrate a GIC in version 1 that is currently not supported
by the A-TF. This change hijacks GICv2 driver to handle the GICv1 as
GICv1 is compatible enough with GICv2 as far as the platform does not
attempt to play with virtualization support or some GICv2 specific power
features. This support mandates a hack in the GICv2 driver
gicv2_driver_init() routine regarding a debug assertion.

Change-Id: Ic2cb3055f1319a83455571d6d918661da583f179
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>