Fix x30 reporting for unhandled exceptions
Crash dumps for unhandled exceptions currently always report the
exception vector address as "x30", because the exception vectors use
branch-with-link to jump to the crash handler (thus overwriting the x30
register from the code that actually caused the exception). This is not
what the reader would expect and also overall not very useful, so this
patch fixes them up to preserve the register correctly.

Change-Id: I2a3636f2943f79bab0cd911f89d070012e697c2a
Signed-off-by: Julius Werner <jwerner@chromium.org>