ARMv7: introduce Cortex-A9
As Cortex-A9 needs to manually enable program flow prediction,
do not reset SCTLR[Z] at entry. Platform should enable it only
once MMU is enabled.

PLAT_OUTER_CACHE shall be enabled if the system includes an outer
cache. If so, the platform is expected to implement an API for
flushing the outer cache. When the API plat_flush_outer_cache() is
called, it is the platform responsibility to check if outer cache is
enabled (and needs a flush maintenance operation).

Change-Id: I34e1ee2da73221903f7767f23bc6fc10ad01e3de
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>