Allow lower ELs to use and access SVE (fixes #520)
Currently booting an SVE enabled kernel through the trusted path ends
up crashing the kernel. To fix I've added a config flag called
ENABLE_SVE_FOR_LOWER_ELS which defaults to 1 and allows the kernel to
access SVE registers. If the trusted firmware ever wants to use SVE
itself additional changes will need to be made.

The changes are based on a similar patch added to the bootwrapper.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Cc: Dave Martin <Dave.Martin@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>